added user assosiation to a comment
When a user is logged in and posts a comment the user auth username is used as the username for the posted comment at the blog details page.
in the fututre i want to make it possible for a non loged in user to post a comment but that user would need to leave some details behind such as an e-mail of some sort.